Addressed race condition between Python and parseEidString()
It appears that parseEidString() temporarily mutates the string passed to it.
It attempts to replace the colon with a null byte in order to create a substring
and pass it as a paramater to a different ION function.

As it turns out, the EID string used to represent a bundle destination in
bp_send was a direct referece to a string managed by the Python interpreter.
Inserting a zero into this while the interpreter is running causes an
"embedded null character" PyError.

In this commit we make a full copy of the EID string in both our BpSAP and our TxInfo,
as to avoid direct references to Python-managed memory.

This prevents null errors on the Python side of things, but it does not address race conditions between
C-level code since we now have even more tasks mutating the EID string (strcpy, parseEidString).

The next commits will attempt to implement mutual exclusion of the EID string, most likely through an SDR transaction.